requirements.txt
# Core RAG Components
langchain==0.1.0
langchain-community==0.0.13
sentence-transformers==2.3.1
faiss-cpu==1.7.4

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3

# LLM
llama-cpp-python==0.2.27
# Or use: transformers, torch for direct LLaMA loading

# UI
streamlit==1.29.0
gradio==4.12.0

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
tqdm==4.66.1
